Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/autofocus/supported-elements-expected.txt


FAIL Contenteditable element should support autofocus assert_equals: expected "DIV" but got "BODY"
FAIL Element with tabindex should support autofocus assert_equals: expected "SPAN" but got "BODY"
PASS Non-HTMLElement should not support autofocus
